Renamed app profile to personal
