+++ personal/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from profile.models import School
+from personal.models import School
 
 
 @admin.register(School)
