+++ personal/fixtures/counties.json
@@ -1,69 +1,69 @@
 [
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 0,
         "fields": {
             "name": "Neur\u010den\u00fd kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 1,
         "fields": {
             "name": "Bratislavsk\u00fd kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 2,
         "fields": {
             "name": "Trnavsk\u00fd kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 3,
         "fields": {
             "name": "Tren\u010diansky kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 4,
         "fields": {
             "name": "Nitriansky kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 5,
         "fields": {
             "name": "\u017dilinsk\u00fd kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 6,
         "fields": {
             "name": "Banskobystrick\u00fd kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 7,
         "fields": {
             "name": "Pre\u0161ovsk\u00fd kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 8,
         "fields": {
             "name": "Ko\u0161ick\u00fd kraj"
         }
     },
     {
-        "model": "profile.County",
+        "model": "personal.County",
         "pk": 9,
         "fields": {
             "name": "Zahrani\u010die"
