+++ personal/fixtures/profiles.json
@@ -1,6 +1,6 @@
 [
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 4,
         "fields": {
             "user": 4,
@@ -12,7 +12,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 5,
         "fields": {
             "user": 47,
@@ -23,7 +23,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 6,
         "fields": {
             "user": 42,
@@ -34,7 +34,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 7,
         "fields": {
             "user": 3,
@@ -45,7 +45,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 8,
         "fields": {
             "user": 5,
@@ -56,7 +56,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 9,
         "fields": {
             "user": 6,
@@ -67,7 +67,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 10,
         "fields": {
             "user": 7,
@@ -78,7 +78,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 11,
         "fields": {
             "user": 8,
@@ -89,7 +89,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 12,
         "fields": {
             "user": 9,
@@ -100,7 +100,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 13,
         "fields": {
             "user": 10,
@@ -111,7 +111,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 14,
         "fields": {
             "user": 11,
@@ -122,7 +122,7 @@
         }
     },
     {
-        "model": "profile.Profile",
+        "model": "personal.Profile",
         "pk": 15,
         "fields": {
             "user": 12,
