+++ personal/fixtures/schools_custom.json
@@ -1,6 +1,6 @@
 [
     {
-        "model": "profile.School",
+        "model": "personal.School",
         "pk": 0,
         "fields": {
             "district": 0,
@@ -8,7 +8,7 @@
         }
     },
     {
-        "model": "profile.School",
+        "model": "personal.School",
         "pk": 1,
         "fields": {
             "district": 0,
