--- profile/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class ProfileConfig(AppConfig):
-    name = 'profile'
-    verbose_name = 'profily'
